% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dummy.R
\name{dummyFunction1}
\alias{dummyFunction1}
\title{Returns the value passed as it is}
\usage{
dummyFunction1(dummyValue)
}
\arguments{
\item{...}{integer}
}
\value{
Exact same value as input parameters
}
\description{
Returns the value passed as it is
}
\author{
Saradindu Sengupta \email{saradindu.mi1@iiitmk.ac.in}
}
