% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dummy.R
\name{dummyFunction2}
\alias{dummyFunction2}
\title{Returns the summed up value of the passed value twice}
\usage{
dummyFunction2(dummyValue)
}
\arguments{
\item{...}{integer}
}
\value{
Summation of the passes value twice
}
\description{
Returns the summed up value of the passed value twice
}
\author{
Saradindu Sengupta \email{saradindu.mi1@iiitmk.ac.in}
}
