% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dummy.R
\name{dummyFunction3}
\alias{dummyFunction3}
\title{Returns the value passed multiplicand by 3}
\usage{
dummyFunction3(dummyValue)
}
\arguments{
\item{...}{integer}
}
\value{
Passed value multiplicand by 3
}
\description{
Returns the value passed multiplicand by 3
}
\author{
Saradindu Sengupta \email{saradindu.mi1@iiitmk.ac.in}
}
