% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dummy.R
\name{dummyFunction4}
\alias{dummyFunction4}
\title{Returns the value passed multiplicand by 7}
\usage{
dummyFunction4(dummyValue)
}
\arguments{
\item{...}{integer}
}
\value{
Passed value multiplicand by 7
}
\description{
Returns the value passed multiplicand by 7
}
\author{
Saradindu Sengupta \email{saradindu.mi1@iiitmk.ac.in}
}
