% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dummy.R
\name{dummyFunction5}
\alias{dummyFunction5}
\title{Returns the value passed divided by 3}
\usage{
dummyFunction5(dummyValue)
}
\arguments{
\item{...}{integer}
}
\value{
Passed value divided by 3
}
\description{
Returns the value passed divided by 3
}
\author{
Saradindu Sengupta \email{saradindu.mi1@iiitmk.ac.in}
}
