 % Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moon_phase.R
\name{is_moon_waxing}
\alias{is_moon_waxing}
\title{Check if the Moon is Waxing or Waning}
\arguments{
\item{jd_ut}{Julian day number in Universal Time (numeric)}
}
\value{
A message indicating whether the Moon is waxing or waning.
}
\description{
Determines whether the Moon is in its waxing or waning phase
based on the Julian day.
}
\examples{
is_moon_waxing(2460760.5)
}
