% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_eclipse_data.R
\name{magnitude_of_sol_eclipse}
\alias{magnitude_of_sol_eclipse}
\title{Compute the Magnitude of a Solar Eclipse}
\usage{
magnitude_of_sol_eclipse(jd, lat, lon, alt)
}
\arguments{
\item{jd}{Julian day number (numeric) representing the date and time.}

\item{lat}{Latitude of the given place (numeric, in degrees).}

\item{lon}{Longitude of the given place (numeric, in degrees).}

\item{alt}{Altitude of the given place (numeric, in meters).}
}
\value{
A list containing eclipse data, including the magnitude.
}
\description{
This function calculates the magnitude of a solar eclipse at a given location
based on the Julian day number, latitude, longitude, and altitude.
}
\examples{
magnitude_of_sol_eclipse(2460760.5, 28.7041, 77.1025, 0)
}
