% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planetary_data.R
\name{planetary_data}
\alias{planetary_data}
\alias{planetary_phase_data}
\title{Retrieve Planetary Phase Data}
\usage{
planetary_phase_data(jd_ut, planet)
}
\arguments{
\item{jd_ut}{Julian day number in Universal Time (numeric).}

\item{planet}{Planet number according to the Swiss Ephemeris (integer).
Example: 0 for Sun, 1 for Moon, 4 for Mars, etc.}
}
\value{
A list containing phase angle, magnitude, and other planetary properties.
}
\description{
Computes and returns phase data for a given planet
using the Swiss Ephemeris.
}
\examples{
planetary_phase_data(2460760.5, 4)  # Retrieves phase data for Mars
}
